+++ laktory/cli/_run.py
@@ -44,6 +44,17 @@ def run(
     filepath: Annotated[
         str, typer.Option(help="Stack (yaml) filepath.")
     ] = "./stack.yaml",
+    var: Annotated[
+        list[str],
+        typer.Option("--var", help="Variable override as key=value. Can be repeated."),
+    ] = [],
+    var_file: Annotated[
+        str,
+        typer.Option(
+            "--var-file",
+            help="YAML file of variable overrides. Auto-discovers variables[.{env}].yaml if not set.",
+        ),
+    ] = None,
 ):
     """
     Execute remote job or DLT pipeline and monitor failures until completion.
@@ -66,11 +77,19 @@ def run(
         Name of the environment.
     filepath:
         Stack (yaml) filepath.
+    var:
+        Variable override as `key=value`. Can be repeated. Overrides variables
+        defined in the stack YAML and in `--var-file`.
+    var_file:
+        Path to a YAML file of variable overrides. If not provided, a
+        `variables[.{env}].yaml` file next to the stack file is used automatically
+        when present.
 
     Examples
     --------
     ```cmd
     laktory run --env dev --dbks-pipeline pl-stock-prices --full_refresh --action CANCEL
+    laktory run --env dev --dbks-job my-job --var profile=MY_PROFILE
     ```
 
     References
@@ -78,7 +97,6 @@ def run(
     * [CLI](https://www.laktory.ai/concepts/cli/)
 
     """
-
     # Set Resource Name
     if databricks_job and databricks_pipeline:
         raise ValueError("Only one of `job` or `dlt` should be set.")
@@ -89,6 +107,8 @@ def run(
     controller = CLIController(
         env=environment,
         stack_filepath=filepath,
+        var_list=var,
+        var_file_path=var_file,
     )
     dispatcher = Dispatcher(stack=controller.stack, env=controller.env)
     dispatcher.get_resource_ids()